refactor: Remove explicit property enumerations from @haiku/core (709220553259038)
- Verified migration is working
- Verified demo behavior is still correct
- Verified in-app editing is working
- Updated a couple of unit tests as appropriate
