+++ packages/haiku-serialization/src/bll/ActiveComponent.js
@@ -4,11 +4,11 @@ const pretty = require('pretty')
 const async = require('async')
 const jss = require('json-stable-stringify')
 const pascalcase = require('pascalcase')
-const {LAYOUT_3D_SCHEMA} = require('@haiku/core/lib/properties/dom/schema')
 const {PlaybackSetting} = require('@haiku/core/lib/HaikuTimeline')
 const {HAIKU_ID_ATTRIBUTE, HAIKU_TITLE_ATTRIBUTE, HAIKU_VAR_ATTRIBUTE} = require('@haiku/core/lib/HaikuElement')
 const {sortedKeyframes} = require('@haiku/core/lib/Transitions').default
 const HaikuComponent = require('@haiku/core/lib/HaikuComponent').default
+const {LAYOUT_3D_SCHEMA} = require('@haiku/core/lib/HaikuComponent')
 const HaikuDOMAdapter = require('@haiku/core/lib/adapters/dom').default
 const {InteractionMode, isPreviewMode} = require('@haiku/core/lib/helpers/interactionModes')
 const Layout3D = require('@haiku/core/lib/Layout3D').default
