+++ packages/haiku-serialization/src/bll/Element.js
@@ -3,9 +3,8 @@ const HaikuElement = require('@haiku/core/lib/HaikuElement').default
 const Layout3D = require('@haiku/core/lib/Layout3D').default
 const cssQueryTree = require('@haiku/core/lib/helpers/cssQueryTree').default
 const composedTransformsToTimelineProperties = require('@haiku/core/lib/helpers/composedTransformsToTimelineProperties').default
-const {LAYOUT_3D_SCHEMA} = require('@haiku/core/lib/properties/dom/schema')
+const {LAYOUT_3D_SCHEMA} = require('@haiku/core/lib/HaikuComponent')
 const KnownDOMEvents = require('@haiku/core/lib/renderers/dom/Events').default
-const DOMSchema = require('@haiku/core/lib/properties/dom/schema').default
 const titlecase = require('titlecase')
 const decamelize = require('decamelize')
 const polygonOverlap = require('polygon-overlap')
@@ -1083,11 +1082,8 @@ class Element extends BaseModel {
   getBuiltinAddressables () {
     const builtinAddressables = {}
 
-    // Start with the basic hardcoded DOM schema; we'll add component-specifics if necessary
-    if (DOMSchema[this.getSafeDomFriendlyName()]) {
-      // This assigns so-called 'cluster' properties if any are deemed such
-      Property.assignDOMSchemaProperties(builtinAddressables, this)
-    }
+    // This assigns so-called 'cluster' properties if any are deemed such
+    Property.assignDOMSchemaProperties(builtinAddressables, this)
 
     return builtinAddressables
   }
